backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
attrs==25.4.0
azure-core==1.34.0
azure-digitaltwins-core==1.2.0
azure-identity==1.23.0
beautifulsoup4==4.14.3
bitarray==3.8.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.5
click==8.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
curl_cffi==0.13.0
cycler==0.12.1
cytoolz==1.1.0
debugpy==1.8.14
decorator==5.2.1
et_xmlfile==2.0.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
executing==2.2.0
fastapi==0.124.0
feedparser==6.0.12
filelock==3.18.0
fonttools==4.58.4
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hexbytes==1.3.1
httplib2==0.31.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
multitasking==0.0.12
narwhals==2.11.0
nest-asyncio==1.6.0
networkx==3.5
npm==0.1.1
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
optional-django==0.1.0
packaging==25.0
pandas==2.3.0
parsimonious==0.10.0
parso==0.8.4
peewee==3.18.3
pillow==11.2.1
platformdirs==4.3.8
plotly==6.5.0
prompt_toolkit==3.0.51
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyngrok==7.4.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyunormalize==17.0.0
python-multipart
PyYAML==6.0.3
pyzmq==27.0.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.4
requests-oauthlib==2.0.0
rlp==4.1.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.2
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
toolz==1.1.0
torch==2.7.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.13
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
yfinance==0.2.66
pypdf
sqlalchemy
psycopg2-binary
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
email-validator
langchain
langchain-google-genai
langchain-community
chromadb
tiktoken
celery
redis
minio
# sentence-transformers REMOVED (replaced by Gemini Embeddings API)
python-magic
asyncpg
redis[asyncio]
anyio[trio]
